feat: Add the soft option to SAML options
When setting up the providr you can now set the soft option.

  provider :saml,
    issuer: 'Example',
    idp_sso_target_url: 'example.com/sso',
    idp_cert: 'mycert',
    soft: true

The default is still false, but you can now over-ride this to turn
off raising of SAML validation errors.

This is useful when testing or in environment where you don't really
want to set up a real SAML server with signed requests etc.