feat: Add `uid_attribute` option to control the attribute used for the user id.
Some SAML 2.0 IdPs use the transient name identifier format. In that case the
name identifier changes for each login, which makes the name identifier
unsuitable as a user identifier.

This patch introduces the `uid_attribute` option, that allows us to use a
SAML 2.0 attribute as the user identifier instead of the name identifier.

Fixes issue #120.